relocate check-type
